fix: remove logs + unused files
