--- partnerThemeData.ts
@@ -1,919 +1,0 @@
-const partnerThemes = {
-  'meta': { 'pagination': { 'page': 1, 'pageSize': 25, 'pageCount': 1, 'total': 13 } }, 'data': [{
-    'id': 13, 'attributes': {
-      'uid': 'WMLseb',
-      'PartnerName': 'WatermelonSeb',
-      'createdAt': '2024-06-18T14:06:20.528Z',
-      'updatedAt': '2024-07-07T22:57:06.527Z',
-      'publishedAt': null,
-      'lightConfig': {
-        'config': {
-          'theme': {
-            'shape': {
-              'borderRadius': 16,
-              'borderRadiusTertiary': 24,
-              'borderRadiusSecondary': 16,
-            },
-            'palette': {
-              'grey': { '200': '#F0E5E8', '300': '#E5D7DA', '700': '#7A666B', '800': '#58373F' },
-              'text': { 'primary': '#190006', 'secondary': '#766066' },
-              'primary': { 'main': '#f7557c' },
-              'secondary': { 'main': '#027944' },
-              'background': { 'paper': '#ffffff', 'default': '#FFEFF3' },
-            },
-            'container': { 'boxShadow': '0px 8px 32px rgba(0, 0, 0, 0.12)', 'borderRadius': '16px' },
-            'components': { 'MuiCard': { 'defaultProps': { 'variant': 'elevation' } } },
-            'typography': 'Inter, sans-serif',
-          }, 'appearance': 'light',
-        },
-        'customization': {
-          'palette': {
-            'accent1': { 'main': '#852c41' },
-            'primary': { 'main': '#ae2848' },
-            'surface1': { 'main': '#FFEFF3' },
-            'surface2': { 'main': '#ffffff' },
-            'secondary': { 'main': '#027944' },
-            'accent1Alt': { 'main': '#ffffff' },
-            'background': { 'main': '#f7557c' },
-          }, 'logoColor': '#FFFFFF',
-        },
-      },
-      'darkConfig': null,
-      'SelectableInMenu': true,
-      'BackgroundColorLight': '#f7557c',
-      'BackgroundColorDark': null,
-      'PartnerURL': null,
-      'BackgroundImageLight': { 'data': null },
-      'BackgroundImageDark': { 'data': null },
-    },
-  }, {
-    'id': 7, 'attributes': {
-      'uid': 'allbridge',
-      'PartnerName': 'Allbridge',
-      'createdAt': '2024-07-01T23:52:01.102Z',
-      'updatedAt': '2024-07-09T12:17:04.279Z',
-      'publishedAt': '2024-07-09T12:17:04.266Z',
-      'lightConfig': null,
-      'darkConfig': {
-        'config': {
-          'theme': {
-            'shape': { 'borderRadius': 20, 'borderRadiusSecondary': 32 },
-            'palette': {
-              'grey': { '800': '#363636' },
-              'text': { 'primary': '#d3d3d3', 'secondary': '#808080' },
-              'primary': { 'main': '#6cf9d8' },
-              'secondary': { 'main': '#c5d5e0' },
-              'background': { 'paper': '#202020', 'default': '#202020' },
-            },
-            'container': { 'boxShadow': '0px 8px 32px rgba(0, 0, 0, 0.08)', 'borderRadius': '16px' },
-            'typography': { 'fontFamily': 'Poppins, sans-serif' },
-          }, 'appearance': 'dark',
-        },
-        'customization': {
-          'palette': {
-            'accent1': { 'main': '#6cf9d8' },
-            'accent2': { 'main': '#5aac8f' },
-            'primary': { 'main': '#6cf9d8' },
-            'surface1': { 'main': '#202020' },
-            'surface2': { 'main': '#202020' },
-            'secondary': { 'main': '#c5d5e0' },
-            'accent1Alt': { 'main': '#6cf9d8' },
-            'background': { 'main': '#000000' },
-          }, 'typography': { 'fontFamily': 'var(--font-sora)' },
-        },
-      },
-      'SelectableInMenu': false,
-      'BackgroundColorLight': null,
-      'BackgroundColorDark': '#000000',
-      'PartnerURL': null,
-      'BackgroundImageLight': { 'data': null },
-      'BackgroundImageDark': { 'data': null },
-    },
-  }, {
-    'id': 11, 'attributes': {
-      'uid': 'symbiosis',
-      'PartnerName': 'Symbiosis',
-      'createdAt': '2024-07-02T07:35:20.765Z',
-      'updatedAt': '2024-07-09T15:38:05.280Z',
-      'publishedAt': '2024-07-09T12:16:40.986Z',
-      'lightConfig': {
-        'config': {
-          'theme': {
-            'shape': {
-              'borderRadius': 12,
-              'borderRadiusTertiary': 24,
-              'borderRadiusSecondary': 12,
-            },
-            'palette': {
-              'grey': { '200': '#F0E5E8', '300': '#E5D7DA', '700': '#7A666B', '800': '#58373F' },
-              'text': { 'primary': '#190006', 'secondary': '#766066' },
-              'primary': { 'main': '#000000' },
-              'secondary': { 'main': '#75fb6e' },
-              'background': { 'paper': '#ffffff', 'default': '#f9f9f9' },
-            },
-            'container': { 'boxShadow': '0px 8px 32px rgba(0, 0, 0, 0.12)', 'borderRadius': '16px' },
-            'components': { 'MuiCard': { 'defaultProps': { 'variant': 'elevation' } } },
-            'typography': { 'fontFamily': 'Inter, sans-serif' },
-          }, 'variant': 'compact', 'appearance': 'light',
-        },
-        'customization': {
-          'palette': {
-            'accent1': { 'main': '#291a40ff' },
-            'accent2': { 'main': '#75fb6eff' },
-            'primary': { 'main': '#000000' },
-            'surface1': { 'main': '#f9f9f9' },
-            'surface2': { 'main': '#f9f9f9' },
-            'secondary': { 'main': '#F5B5FF' },
-            'accent1Alt': { 'main': '#291a40ff' },
-            'background': { 'default': '#f9f9f9' },
-          }, 'typography': '__Sora_ec8806',
-        },
-      },
-      'darkConfig': null,
-      'SelectableInMenu': false,
-      'BackgroundColorLight': '#f9f9f9',
-      'BackgroundColorDark': null,
-      'PartnerURL': null,
-      'BackgroundImageLight': {
-        'data': {
-          'id': 451, 'attributes': {
-            'name': 'symbiosis-bg.png',
-            'alternativeText': null,
-            'caption': null,
-            'width': 1440,
-            'height': 950,
-            'formats': {
-              'large': {
-                'ext': '.png',
-                'url': '/uploads/large_symbiosis_bg_0bdf8625fb.png',
-                'hash': 'large_symbiosis_bg_0bdf8625fb',
-                'mime': 'image/png',
-                'name': 'large_symbiosis-bg.png',
-                'path': null,
-                'size': 20.87,
-                'width': 1000,
-                'height': 660,
-                'sizeInBytes': 20865,
-              },
-              'small': {
-                'ext': '.png',
-                'url': '/uploads/small_symbiosis_bg_0bdf8625fb.png',
-                'hash': 'small_symbiosis_bg_0bdf8625fb',
-                'mime': 'image/png',
-                'name': 'small_symbiosis-bg.png',
-                'path': null,
-                'size': 8.64,
-                'width': 500,
-                'height': 330,
-                'sizeInBytes': 8642,
-              },
-              'medium': {
-                'ext': '.png',
-                'url': '/uploads/medium_symbiosis_bg_0bdf8625fb.png',
-                'hash': 'medium_symbiosis_bg_0bdf8625fb',
-                'mime': 'image/png',
-                'name': 'medium_symbiosis-bg.png',
-                'path': null,
-                'size': 14.75,
-                'width': 750,
-                'height': 495,
-                'sizeInBytes': 14746,
-              },
-              'thumbnail': {
-                'ext': '.png',
-                'url': '/uploads/thumbnail_symbiosis_bg_0bdf8625fb.png',
-                'hash': 'thumbnail_symbiosis_bg_0bdf8625fb',
-                'mime': 'image/png',
-                'name': 'thumbnail_symbiosis-bg.png',
-                'path': null,
-                'size': 3.54,
-                'width': 236,
-                'height': 156,
-                'sizeInBytes': 3540,
-              },
-            },
-            'hash': 'symbiosis_bg_0bdf8625fb',
-            'ext': '.png',
-            'mime': 'image/png',
-            'size': 23.43,
-            'url': '/uploads/symbiosis_bg_0bdf8625fb.png',
-            'previewUrl': null,
-            'provider': 'local',
-            'provider_metadata': null,
-            'createdAt': '2024-07-08T14:36:11.058Z',
-            'updatedAt': '2024-07-08T14:36:11.058Z',
-          },
-        },
-      },
-      'BackgroundImageDark': { 'data': null },
-    },
-  }, {
-    'id': 2, 'attributes': {
-      'uid': 'WML',
-      'PartnerName': 'Watermelon',
-      'createdAt': '2024-06-18T14:06:20.528Z',
-      'updatedAt': '2024-07-07T22:57:06.527Z',
-      'publishedAt': null,
-      'lightConfig': {
-        'config': {
-          'theme': {
-            'shape': {
-              'borderRadius': 16,
-              'borderRadiusTertiary': 24,
-              'borderRadiusSecondary': 16,
-            },
-            'palette': {
-              'grey': { '200': '#F0E5E8', '300': '#E5D7DA', '700': '#7A666B', '800': '#58373F' },
-              'text': { 'primary': '#190006', 'secondary': '#766066' },
-              'primary': { 'main': '#f7557c' },
-              'secondary': { 'main': '#027944' },
-              'background': { 'paper': '#ffffff', 'default': '#FFEFF3' },
-            },
-            'container': { 'boxShadow': '0px 8px 32px rgba(0, 0, 0, 0.12)', 'borderRadius': '16px' },
-            'components': { 'MuiCard': { 'defaultProps': { 'variant': 'elevation' } } },
-            'typography': 'Inter, sans-serif',
-          }, 'appearance': 'light',
-        },
-        'customization': {
-          'palette': {
-            'accent1': { 'main': '#852c41' },
-            'primary': { 'main': '#ae2848' },
-            'surface1': { 'main': '#FFEFF3' },
-            'surface2': { 'main': '#ffffff' },
-            'secondary': { 'main': '#027944' },
-            'accent1Alt': { 'main': '#ffffff' },
-            'background': { 'main': '#f7557c' },
-          }, 'logoColor': '#FFFFFF',
-        },
-      },
-      'darkConfig': null,
-      'SelectableInMenu': true,
-      'BackgroundColorLight': '#f7557c',
-      'BackgroundColorDark': null,
-      'PartnerURL': null,
-      'BackgroundImageLight': { 'data': null },
-      'BackgroundImageDark': { 'data': null },
-    },
-  }, {
-    'id': 1, 'attributes': {
-      'uid': 'w95',
-      'PartnerName': 'Windows 95',
-      'createdAt': '2024-06-11T15:41:09.062Z',
-      'updatedAt': '2024-07-04T14:31:25.475Z',
-      'publishedAt': null,
-      'lightConfig': {
-        'config': {
-          'theme': {
-            'shape': { 'borderRadius': 0, 'borderRadiusTertiary': 0, 'borderRadiusSecondary': 0 },
-            'palette': {
-              'grey': { '300': '#bbbbbb' },
-              'text': { 'primary': '#30313d', 'secondary': '#6d6e78' },
-              'primary': { 'main': '#0000ff' },
-              'secondary': { 'main': '#f7c303' },
-              'background': { 'paper': '#dfdfdf', 'default': '#c0c9d2' },
-            },
-            'container': { 'border': '1px solid #0a0a0a' },
-            'components': {
-              'MuiCard': {
-                'variants': [{ 'props': { 'variant': 'outlined' }, 'style': { 'borderWidth': '0' } }],
-                'defaultProps': { 'variant': 'outlined' },
-                'styleOverrides': {
-                  'root': {
-                    '&:hover': { 'cursor': 'pointer', 'backgroundColor': '#cccccc' },
-                    'boxShadow': 'inset -1px -1px #0a0a0a, inset 1px 1px #ffffff, inset -2px -2px #808080, inset 2px 2px #dfdfdf',
-                    'borderWidth': '0',
-                  },
-                },
-              },
-              'MuiButton': {
-                'styleOverrides': {
-                  'root': {
-                    '&:hover': { 'boxShadow': 'inset -1px -1px #0a0a0a, inset 1px 1px #ffffff, inset -2px -2px #808080, inset 2px 2px #dfdfdf' },
-                    'boxShadow': 'inset -1px -1px #0a0a0a, inset 1px 1px #ffffff, inset -2px -2px #808080, inset 2px 2px #dfdfdf',
-                  },
-                },
-              },
-              'MuiInputCard': {
-                'styleOverrides': {
-                  'root': {
-                    'boxShadow': 'inset -1px -1px #ffffff, inset 1px 1px #0a0a0a, inset -2px -2px #dfdfdf, inset 2px 2px #808080',
-                    'backgroundColor': '#ffffff',
-                  },
-                },
-              },
-              'MuiIconButton': {
-                'styleOverrides': {
-                  'root': {
-                    '&:hover': {
-                      'color': 'inherit',
-                      'backgroundColor': '#cccccc',
-                    },
-                    'boxShadow': 'inset -1px -1px #0a0a0a, inset 1px 1px #ffffff, inset -2px -2px #808080, inset 2px 2px #dfdfdf',
-                    'borderRadius': '0',
-                    'backgroundColor': '#dfdfdf',
-                  },
-                },
-              },
-            },
-            'typography': 'Inter',
-          },
-        },
-        'customization': {
-          'palette': {
-            'text': { 'primary': '#30313d', 'secondary': '#6d6e78' },
-            'accent1': { 'main': '#30313d' },
-            'accent2': { 'main': '#f7c303' },
-            'primary': { 'main': '#0000ff' },
-            'surface1': { 'main': '#c0c9d2' },
-            'secondary': { 'main': '#f7c303' },
-            'accent1Alt': { 'main': '#30313d' },
-            'background': { 'main': '#008080' },
-          },
-        },
-      },
-      'darkConfig': null,
-      'SelectableInMenu': true,
-      'BackgroundColorLight': '#008080',
-      'BackgroundColorDark': null,
-      'PartnerURL': null,
-      'BackgroundImageLight': { 'data': null },
-      'BackgroundImageDark': { 'data': null },
-    },
-  }, {
-    'id': 8, 'attributes': {
-      'uid': 'cbridge',
-      'PartnerName': 'cBridge',
-      'createdAt': '2024-07-02T00:03:23.216Z',
-      'updatedAt': '2024-07-09T12:16:36.976Z',
-      'publishedAt': '2024-07-09T12:16:36.972Z',
-      'lightConfig': null,
-      'darkConfig': {
-        'config': {
-          'theme': {
-            'shape': { 'borderRadius': 16, 'borderRadiusSecondary': 16 },
-            'palette': {
-              'grey': { '700': '#5c5c5c', '800': '#222530' },
-              'text': { 'secondary': '#8f9bb3' },
-              'primary': { 'main': '#3366ff' },
-              'secondary': { 'main': '#8fb4ff' },
-              'background': { 'paper': '#17171a', 'default': '#222530' },
-            },
-            'container': { 'boxShadow': '0px 8px 32px rgba(0, 0, 0, 0.08)', 'borderRadius': '16px' },
-            'typography': { 'fontFamily': 'Inter, sans-serif' },
-          }, 'variant': 'wide', 'appearance': 'dark',
-        },
-        'customization': {
-          'palette': {
-            'accent1': { 'main': '#ffffff' },
-            'accent2': { 'main': '#8e9bb3ff' },
-            'primary': { 'main': '#3366ff' },
-            'surface1': { 'main': '#222530' },
-            'surface2': { 'main': '#222530' },
-            'secondary': { 'main': '#8fb4ff' },
-            'accent1Alt': { 'main': '#ffffff' },
-            'background': { 'main': '#000000' },
-          }, 'typography': 'Inter, sans-serif',
-        },
-      },
-      'SelectableInMenu': false,
-      'BackgroundColorLight': null,
-      'BackgroundColorDark': '#000000',
-      'PartnerURL': null,
-      'BackgroundImageLight': { 'data': null },
-      'BackgroundImageDark': { 'data': null },
-    },
-  }, {
-    'id': 4, 'attributes': {
-      'uid': 'across',
-      'PartnerName': 'Across',
-      'createdAt': '2024-06-11T15:41:09.062Z',
-      'updatedAt': '2024-07-09T12:16:31.221Z',
-      'publishedAt': '2024-07-09T12:16:31.217Z',
-      'lightConfig': null,
-      'darkConfig': {
-        'config': {
-          'theme': {
-            'shape': { 'borderRadius': 12, 'borderRadiusSecondary': 12 },
-            'palette': {
-              'grey': { '200': '#f7f0e9', '300': '#f4ebe1', '800': '#414247' },
-              'text': { 'primary': '#c5d5e0', 'secondary': '#787c82' },
-              'primary': { 'main': '#6cf9d8' },
-              'secondary': { 'main': '#c5d5e0' },
-              'background': { 'paper': '#393a3f', 'default': '#34353a' },
-            },
-            'container': { 'boxShadow': '0px 8px 32px rgba(0, 0, 0, 0.08)', 'borderRadius': '16px' },
-            'typography': { 'fontFamily': 'Barlow, Inter, sans-serif' },
-          }, 'variant': 'compact', 'appearance': 'dark', 'subvariant': 'default',
-        },
-        'customization': {
-          'palette': {
-            'text': { 'primary': '#30313d', 'secondary': '#6d6e78' },
-            'accent1': { 'main': '#ffffff' },
-            'accent2': { 'main': '#ffffff' },
-            'primary': { 'main': '#6cf9d8' },
-            'surface1': { 'main': '#34353a' },
-            'secondary': { 'main': '#c5d5e0' },
-            'accent1Alt': { 'main': '#ffffff' },
-            'background': { 'main': '#34353a' },
-          },
-        },
-      },
-      'SelectableInMenu': false,
-      'BackgroundColorLight': null,
-      'BackgroundColorDark': '#2d2e33ff',
-      'PartnerURL': null,
-      'BackgroundImageLight': { 'data': null },
-      'BackgroundImageDark': { 'data': null },
-    },
-  }, {
-    'id': 6, 'attributes': {
-      'uid': 'stargate',
-      'PartnerName': 'Stargate',
-      'createdAt': '2024-07-01T22:09:09.924Z',
-      'updatedAt': '2024-07-09T12:16:44.245Z',
-      'publishedAt': '2024-07-09T12:16:44.241Z',
-      'lightConfig': null,
-      'darkConfig': {
-        'config': {
-          'theme': {
-            'shape': { 'borderRadius': 12, 'borderRadiusSecondary': 12 },
-            'palette': {
-              'grey': { '800': '#323232' },
-              'primary': { 'main': '#ffffff' },
-              'secondary': { 'main': '#50beaf' },
-              'background': { 'paper': '#1a1a1a', 'default': '#232323' },
-            },
-            'container': { 'boxShadow': '0px 8px 32px rgba(0, 0, 0, 0.08)', 'borderRadius': '16px' },
-            'typography': { 'fontFamily': 'Inter, sans-serif' },
-          }, 'variant': 'compact', 'appearance': 'dark',
-        },
-        'customization': {
-          'palette': {
-            'accent1': { 'main': '#ffffff' },
-            'accent2': { 'main': '#999999' },
-            'primary': { 'main': '#9DA1A3' },
-            'surface1': { 'main': '#323232' },
-            'surface2': { 'main': '#1a1a1a' },
-            'secondary': { 'main': '#50beaf' },
-            'accent1Alt': { 'main': '#ffffff' },
-            'background': { 'main': '#000000' },
-          }, 'typography': '__Sora_ec8806',
-        },
-      },
-      'SelectableInMenu': false,
-      'BackgroundColorLight': null,
-      'BackgroundColorDark': '#000000',
-      'PartnerURL': null,
-      'BackgroundImageLight': { 'data': null },
-      'BackgroundImageDark': {
-        'data': {
-          'id': 286,
-          'attributes': {
-            'name': 'stargate_stars_bg.svg',
-            'alternativeText': null,
-            'caption': null,
-            'width': 1440,
-            'height': 800,
-            'formats': null,
-            'hash': 'stargate_stars_bg_1f8856be9e',
-            'ext': '.svg',
-            'mime': 'image/svg+xml',
-            'size': 854.54,
-            'url': '/uploads/stargate_stars_bg_1f8856be9e.svg',
-            'previewUrl': null,
-            'provider': 'local',
-            'provider_metadata': null,
-            'createdAt': '2024-07-01T22:13:13.744Z',
-            'updatedAt': '2024-07-01T22:13:13.744Z',
-          },
-        },
-      },
-    },
-  }, {
-    'id': 3, 'attributes': {
-      'uid': 'OP',
-      'PartnerName': 'Optimism',
-      'createdAt': '2024-06-20T14:39:00.144Z',
-      'updatedAt': '2024-07-09T22:40:07.732Z',
-      'publishedAt': '2024-06-21T08:43:39.565Z',
-      'lightConfig': {
-        'config': {
-          'theme': {
-            'shape': {
-              'borderRadius': 12,
-              'borderRadiusTertiary': 24,
-              'borderRadiusSecondary': 16,
-            },
-            'palette': {
-              'grey': { '300': '#fff0ca' },
-              'primary': { 'main': '#ff0420' },
-              'secondary': { 'dark': '#F5B5FF', 'main': '#F5B5FF', 'light': '#F5B5FF' },
-              'background': { 'paper': '#ffffff', 'default': '#fdfbef' },
-            },
-            'container': { 'boxShadow': '0px 8px 32px rgba(0, 0, 0, 0.08)', 'borderRadius': '16px' },
-            'typography': { 'fontFamily': '__Sora_ec8806, Inter, sans-serif' },
-          }, 'variant': 'compact', 'appearance': 'light',
-        },
-        'customization': {
-          'palette': {
-            'accent1': { 'main': '#000000' },
-            'accent2': { 'main': '#000000' },
-            'primary': { 'main': '#ff0420' },
-            'surface1': { 'main': '#fdfbef' },
-            'secondary': { 'main': '#F5B5FF' },
-            'accent1Alt': { 'main': '#000000' },
-            'background': { 'main': '#b7e8fb' },
-          }, 'typography': '__Sora_ec8806',
-        },
-      },
-      'darkConfig': null,
-      'SelectableInMenu': true,
-      'BackgroundColorLight': '#b7e8fb',
-      'BackgroundColorDark': null,
-      'PartnerURL': 'https://www.superchain.eco/',
-      'BackgroundImageLight': {
-        'data': {
-          'id': 275,
-          'attributes': {
-            'name': 'Superfest OP.svg',
-            'alternativeText': null,
-            'caption': null,
-            'width': 960,
-            'height': 640,
-            'formats': null,
-            'hash': 'Superfest_OP_6b116a3b5b',
-            'ext': '.svg',
-            'mime': 'image/svg+xml',
-            'size': 7.36,
-            'url': '/uploads/Superfest_OP_6b116a3b5b.svg',
-            'previewUrl': null,
-            'provider': 'local',
-            'provider_metadata': null,
-            'createdAt': '2024-06-24T12:15:04.367Z',
-            'updatedAt': '2024-06-24T12:15:04.367Z',
-          },
-        },
-      },
-      'BackgroundImageDark': { 'data': null },
-    },
-  }, {
-    'id': 5, 'attributes': {
-      'uid': 'hop',
-      'PartnerName': 'Hop',
-      'createdAt': '2024-07-01T10:17:41.039Z',
-      'updatedAt': '2024-07-08T21:28:58.512Z',
-      'publishedAt': '2024-07-01T11:46:44.325Z',
-      'lightConfig': null,
-      'darkConfig': {
-        'config': {
-          'theme': {
-            'shape': { 'borderRadius': 12, 'borderRadiusSecondary': 128 },
-            'palette': {
-              'grey': { '800': '#302b3d' },
-              'text': { 'primary': '#c5d5e0', 'secondary': '#787c82' },
-              'primary': { 'main': '#b23eff' },
-              'secondary': { 'main': '#e5b8ff' },
-              'background': { 'paper': '#272332', 'default': '#302b3d' },
-            },
-            'container': { 'boxShadow': '0px 8px 32px rgba(0, 0, 0, 0.08)', 'borderRadius': '16px' },
-            'typography': { 'fontFamily': 'Inter, sans-serif' },
-          }, 'variant': 'compact', 'appearance': 'dark',
-        },
-        'customization': {
-          'palette': {
-            'accent1': { 'main': '#FFFFFF' },
-            'accent2': { 'main': '#b23eff' },
-            'primary': { 'main': '#b23eff' },
-            'surface1': { 'main': '#302b3d' },
-            'secondary': { 'main': '#e5b8ff' },
-            'accent1Alt': { 'main': '#FFFFFF' },
-            'background': { 'main': '#302b3d' },
-          },
-        },
-      },
-      'SelectableInMenu': false,
-      'BackgroundColorLight': null,
-      'BackgroundColorDark': '#272332',
-      'PartnerURL': null,
-      'BackgroundImageLight': { 'data': null },
-      'BackgroundImageDark': {
-        'data': {
-          'id': 285,
-          'attributes': {
-            'name': 'Hop circles-bg-dark.svg',
-            'alternativeText': null,
-            'caption': null,
-            'width': 1770,
-            'height': 1471,
-            'formats': null,
-            'hash': 'Hop_circles_bg_dark_1dc943c925',
-            'ext': '.svg',
-            'mime': 'image/svg+xml',
-            'size': 1.19,
-            'url': '/uploads/Hop_circles_bg_dark_1dc943c925.svg',
-            'previewUrl': null,
-            'provider': 'local',
-            'provider_metadata': null,
-            'createdAt': '2024-07-01T10:21:28.457Z',
-            'updatedAt': '2024-07-01T10:21:28.457Z',
-          },
-        },
-      },
-    },
-  }, {
-    'id': 10, 'attributes': {
-      'uid': 'hyphen',
-      'PartnerName': 'Hyphen',
-      'createdAt': '2024-07-02T05:46:43.470Z',
-      'updatedAt': '2024-07-09T12:16:49.452Z',
-      'publishedAt': '2024-07-09T12:16:49.447Z',
-      'lightConfig': null,
-      'darkConfig': {
-        'config': {
-          'theme': {
-            'shape': { 'borderRadius': 12, 'borderRadiusSecondary': 16 },
-            'palette': {
-              'grey': { '700': '#707070', '800': '#525252' },
-              'text': { 'primary': '#545757', 'secondary': '#8d9191' },
-              'primary': { 'main': '#615ccd' },
-              'secondary': { 'main': '#8880ff' },
-              'background': { 'paper': '#FFFFFF', 'default': '#FFFFFF' },
-            },
-            'container': { 'boxShadow': '0px 8px 32px rgba(0, 0, 0, 0.08)', 'borderRadius': '16px' },
-            'typography': { 'fontFamily': 'Inter, sans-serif' },
-          }, 'variant': 'compact', 'appearance': 'light',
-        },
-        'customization': {
-          'palette': {
-            'accent1': { 'main': '#ffffff' },
-            'accent2': { 'main': '#615ccd' },
-            'primary': { 'main': '#615ccd' },
-            'surface1': { 'main': '#1f1e1e' },
-            'surface2': { 'main': '#1f1e1e' },
-            'secondary': { 'main': '#8880ff' },
-            'accent1Alt': { 'main': '#ffffff' },
-            'background': { 'main': '#000000' },
-          }, 'typography': 'Inter, sans-serif',
-        },
-      },
-      'SelectableInMenu': false,
-      'BackgroundColorLight': null,
-      'BackgroundColorDark': null,
-      'PartnerURL': null,
-      'BackgroundImageLight': {
-        'data': {
-          'id': 291, 'attributes': {
-            'name': 'Hyphen Bg.jpg',
-            'alternativeText': null,
-            'caption': null,
-            'width': 1440,
-            'height': 900,
-            'formats': {
-              'large': {
-                'ext': '.jpg',
-                'url': '/uploads/large_Hyphen_Bg_6cc6b6d306.jpg',
-                'hash': 'large_Hyphen_Bg_6cc6b6d306',
-                'mime': 'image/jpeg',
-                'name': 'large_Hyphen Bg.jpg',
-                'path': null,
-                'size': 25.22,
-                'width': 1000,
-                'height': 625,
-                'sizeInBytes': 25218,
-              },
-              'small': {
-                'ext': '.jpg',
-                'url': '/uploads/small_Hyphen_Bg_6cc6b6d306.jpg',
-                'hash': 'small_Hyphen_Bg_6cc6b6d306',
-                'mime': 'image/jpeg',
-                'name': 'small_Hyphen Bg.jpg',
-                'path': null,
-                'size': 8.7,
-                'width': 500,
-                'height': 313,
-                'sizeInBytes': 8703,
-              },
-              'medium': {
-                'ext': '.jpg',
-                'url': '/uploads/medium_Hyphen_Bg_6cc6b6d306.jpg',
-                'hash': 'medium_Hyphen_Bg_6cc6b6d306',
-                'mime': 'image/jpeg',
-                'name': 'medium_Hyphen Bg.jpg',
-                'path': null,
-                'size': 16.39,
-                'width': 750,
-                'height': 469,
-                'sizeInBytes': 16390,
-              },
-              'thumbnail': {
-                'ext': '.jpg',
-                'url': '/uploads/thumbnail_Hyphen_Bg_6cc6b6d306.jpg',
-                'hash': 'thumbnail_Hyphen_Bg_6cc6b6d306',
-                'mime': 'image/jpeg',
-                'name': 'thumbnail_Hyphen Bg.jpg',
-                'path': null,
-                'size': 2.91,
-                'width': 245,
-                'height': 153,
-                'sizeInBytes': 2908,
-              },
-            },
-            'hash': 'Hyphen_Bg_6cc6b6d306',
-            'ext': '.jpg',
-            'mime': 'image/jpeg',
-            'size': 377.67,
-            'url': '/uploads/Hyphen_Bg_6cc6b6d306.jpg',
-            'previewUrl': null,
-            'provider': 'local',
-            'provider_metadata': null,
-            'createdAt': '2024-07-02T05:45:39.878Z',
-            'updatedAt': '2024-07-02T05:52:20.860Z',
-          },
-        },
-      },
-      'BackgroundImageDark': {
-        'data': {
-          'id': 291, 'attributes': {
-            'name': 'Hyphen Bg.jpg',
-            'alternativeText': null,
-            'caption': null,
-            'width': 1440,
-            'height': 900,
-            'formats': {
-              'large': {
-                'ext': '.jpg',
-                'url': '/uploads/large_Hyphen_Bg_6cc6b6d306.jpg',
-                'hash': 'large_Hyphen_Bg_6cc6b6d306',
-                'mime': 'image/jpeg',
-                'name': 'large_Hyphen Bg.jpg',
-                'path': null,
-                'size': 25.22,
-                'width': 1000,
-                'height': 625,
-                'sizeInBytes': 25218,
-              },
-              'small': {
-                'ext': '.jpg',
-                'url': '/uploads/small_Hyphen_Bg_6cc6b6d306.jpg',
-                'hash': 'small_Hyphen_Bg_6cc6b6d306',
-                'mime': 'image/jpeg',
-                'name': 'small_Hyphen Bg.jpg',
-                'path': null,
-                'size': 8.7,
-                'width': 500,
-                'height': 313,
-                'sizeInBytes': 8703,
-              },
-              'medium': {
-                'ext': '.jpg',
-                'url': '/uploads/medium_Hyphen_Bg_6cc6b6d306.jpg',
-                'hash': 'medium_Hyphen_Bg_6cc6b6d306',
-                'mime': 'image/jpeg',
-                'name': 'medium_Hyphen Bg.jpg',
-                'path': null,
-                'size': 16.39,
-                'width': 750,
-                'height': 469,
-                'sizeInBytes': 16390,
-              },
-              'thumbnail': {
-                'ext': '.jpg',
-                'url': '/uploads/thumbnail_Hyphen_Bg_6cc6b6d306.jpg',
-                'hash': 'thumbnail_Hyphen_Bg_6cc6b6d306',
-                'mime': 'image/jpeg',
-                'name': 'thumbnail_Hyphen Bg.jpg',
-                'path': null,
-                'size': 2.91,
-                'width': 245,
-                'height': 153,
-                'sizeInBytes': 2908,
-              },
-            },
-            'hash': 'Hyphen_Bg_6cc6b6d306',
-            'ext': '.jpg',
-            'mime': 'image/jpeg',
-            'size': 377.67,
-            'url': '/uploads/Hyphen_Bg_6cc6b6d306.jpg',
-            'previewUrl': null,
-            'provider': 'local',
-            'provider_metadata': null,
-            'createdAt': '2024-07-02T05:45:39.878Z',
-            'updatedAt': '2024-07-02T05:52:20.860Z',
-          },
-        },
-      },
-    },
-  }, {
-    'id': 9, 'attributes': {
-      'uid': 'amarok',
-      'PartnerName': 'Connext',
-      'createdAt': '2024-07-02T00:12:59.242Z',
-      'updatedAt': '2024-07-09T12:16:53.760Z',
-      'publishedAt': '2024-07-09T12:16:53.755Z',
-      'lightConfig': null,
-      'darkConfig': {
-        'config': {
-          'theme': {
-            'shape': { 'borderRadius': 4, 'borderRadiusSecondary': 4 },
-            'palette': {
-              'grey': { '800': '#303030' },
-              'text': { 'primary': '#d3d3d3', 'secondary': '#808080' },
-              'primary': { 'main': '#2563eb' },
-              'secondary': { 'main': '#8fb4ff' },
-              'background': { 'paper': '#1f1e1e', 'default': '#1f1e1e' },
-            },
-            'container': { 'boxShadow': '0px 8px 32px rgba(0, 0, 0, 0.08)', 'borderRadius': '4px' },
-            'typography': { 'fontFamily': 'manrope, sans-serif' },
-          }, 'variant': 'compact', 'appearance': 'dark',
-        },
-        'customization': {
-          'palette': {
-            'accent1': { 'main': '#ffffff' },
-            'accent2': { 'main': '#2563eb' },
-            'primary': { 'main': '#2563eb' },
-            'surface1': { 'main': '#1f1e1e' },
-            'surface2': { 'main': '#1f1e1e' },
-            'secondary': { 'main': '#8fb4ff' },
-            'accent1Alt': { 'main': '#ffffff' },
-            'background': { 'main': '#000000' },
-          }, 'typography': 'manrope',
-        },
-      },
-      'SelectableInMenu': false,
-      'BackgroundColorLight': null,
-      'BackgroundColorDark': '#000000',
-      'PartnerURL': null,
-      'BackgroundImageLight': { 'data': null },
-      'BackgroundImageDark': { 'data': null },
-    },
-  }, {
-    'id': 12, 'attributes': {
-      'uid': 'op2testseb',
-      'PartnerName': 'Optimism',
-      'createdAt': '2024-06-20T14:39:00.144Z',
-      'updatedAt': '2024-07-04T17:11:17.239Z',
-      'publishedAt': null,
-      'lightConfig': {
-        'config': {
-          'theme': {
-            'shape': {
-              'borderRadius': 12,
-              'borderRadiusTertiary': 24,
-              'borderRadiusSecondary': 16,
-            },
-            'palette': {
-              'grey': { '300': '#fff0ca' },
-              'primary': { 'main': '#ff0420' },
-              'secondary': { 'main': '#F5B5FF' },
-              'background': { 'paper': '#ffffff', 'default': '#fdfbef' },
-            },
-            'container': { 'boxShadow': '0px 8px 32px rgba(0, 0, 0, 0.08)', 'borderRadius': '16px' },
-            'typography': { 'fontFamily': '__Sora_ec8806, Inter, sans-serif' },
-          }, 'appearance': 'light',
-        },
-        'customization': {
-          'palette': {
-            'accent1': { 'main': '#000000' },
-            'accent2': { 'main': '#000000' },
-            'primary': { 'main': '#ff0420' },
-            'surface1': { 'main': '#fdfbef' },
-            'secondary': { 'main': '#F5B5FF' },
-            'accent1Alt': { 'main': '#000000' },
-            'background': { 'main': '#b7e8fb' },
-          }, 'typography': { 'fontFamily': 'var(--font-sora)' },
-        },
-      },
-      'darkConfig': null,
-      'SelectableInMenu': false,
-      'BackgroundColorLight': '#b7e8fb',
-      'BackgroundColorDark': null,
-      'PartnerURL': 'https://li.fi',
-      'BackgroundImageLight': {
-        'data': {
-          'id': 275,
-          'attributes': {
-            'name': 'Superfest OP.svg',
-            'alternativeText': null,
-            'caption': null,
-            'width': 960,
-            'height': 640,
-            'formats': null,
-            'hash': 'Superfest_OP_6b116a3b5b',
-            'ext': '.svg',
-            'mime': 'image/svg+xml',
-            'size': 7.36,
-            'url': '/uploads/Superfest_OP_6b116a3b5b.svg',
-            'previewUrl': null,
-            'provider': 'local',
-            'provider_metadata': null,
-            'createdAt': '2024-06-24T12:15:04.367Z',
-            'updatedAt': '2024-06-24T12:15:04.367Z',
-          },
-        },
-      },
-      'BackgroundImageDark': { 'data': null },
-    },
-  }], 'url': 'https://strapi.li.finance',
-};
-
-export default partnerThemes;
