+++ src/app/[lng]/error.tsx
@@ -1,6 +1,8 @@
 'use client'; // Error components must be Client Components
 import dynamic from 'next/dynamic';
 import { useEffect } from 'react';
+import { ThemeProvider } from 'next-themes';
+import { ThemeProviderV2 } from '@/providers/ThemeProviderV2';
 
 const ErrorPage = dynamic(() => import('../ui/error/ErrorPage'), {
   ssr: false,
@@ -18,5 +20,16 @@ export default function Error({
     console.error(error);
   }, [error]);
 
-  return <ErrorPage reset={reset} />;
+  return (
+    <ThemeProvider
+      themes={['dark', 'light']}
+      forcedTheme={'light'}
+      enableSystem
+      enableColorScheme
+    >
+      <ThemeProviderV2 themes={[]}>
+        <ErrorPage reset={reset} />
+      </ThemeProviderV2>
+    </ThemeProvider>
+  );
 }
