+++ src/app/not-found.tsx
@@ -4,7 +4,6 @@ import { Logo } from '@/components/Navbar/Logo/Logo';
 import { NavbarContainer } from '@/components/Navbar/Navbar.style';
 import { NavbarButtons } from '@/components/Navbar/NavbarButtons';
 import { NotFoundComponent } from '@/components/NotFound/NotFound';
-import { PoweredBy } from '@/components/PoweredBy';
 import TranslationsProvider from '@/providers/TranslationProvider';
 import { Link } from '@mui/material';
 import { AppRouterCacheProvider } from '@mui/material-nextjs/v14-appRouter';
@@ -12,9 +11,10 @@ import { cookies } from 'next/headers';
 import RouterLink from 'next/link';
 import { defaultNS, namespaces } from 'src/i18n';
 import Background from '@/components/Background';
+import { ThemeProvider } from 'next-themes';
+import { ThemeProviderV2 } from '@/providers/ThemeProviderV2';
 
 export default async function NotFound() {
-  const { activeThemeMode } = getCookies();
   const cookiesHandler = cookies();
   const locale = cookiesHandler.get('NEXT_LOCALE')?.value ?? 'en';
 
@@ -23,23 +23,29 @@ export default async function NotFound() {
   return (
     <>
       <AppRouterCacheProvider>
-        {/*<ThemeProvider themeMode={activeThemeMode} activeTheme="jumper">*/}
-        <TranslationsProvider
-          namespaces={[defaultNS]}
-          locale={locale}
-          resources={resources}
+        <ThemeProvider
+          themes={['dark', 'light']}
+          forcedTheme={'light'}
+          enableSystem
+          enableColorScheme
         >
-          <Background />
-
-          <NavbarContainer>
-            <Link component={RouterLink} href="/">
-              <Logo variant="default" />
-            </Link>
-            <NavbarButtons />
-          </NavbarContainer>
-          <NotFoundComponent />
-        </TranslationsProvider>
-        {/*</ThemeProvider>*/}
+          <ThemeProviderV2 themes={[]}>
+            <TranslationsProvider
+              namespaces={[defaultNS]}
+              locale={locale}
+              resources={resources}
+            >
+              <Background />
+              <NavbarContainer>
+                <Link component={RouterLink} href="/">
+                  <Logo variant="default" />
+                </Link>
+                <NavbarButtons />
+              </NavbarContainer>
+              <NotFoundComponent />
+            </TranslationsProvider>
+          </ThemeProviderV2>
+        </ThemeProvider>
       </AppRouterCacheProvider>
     </>
   );
