+++ src/hooks/useSuperfest.ts
@@ -1,13 +1,14 @@
-import { useSelectedLayoutSegment } from 'next/navigation';
+import { usePathname } from 'next/navigation';
+import { JUMPER_FEST_PATH } from 'src/const/urls';
 
 interface useSuperfestProps {
   isSuperfest: boolean;
 }
 
 export const useSuperfest = (): useSuperfestProps => {
-  const segment = useSelectedLayoutSegment();
+  const pathname = usePathname();
 
   return {
-    isSuperfest: segment === 'superfest',
+    isSuperfest: pathname?.includes(JUMPER_FEST_PATH) ?? false,
   };
 };
